+++ src/utils/copyToClipboard.ts
@@ -1,0 +1,9 @@
+
+export const copyToClipboard = (text: string) => {
+  const textArea = document.createElement('textarea');
+  document.body.appendChild(textArea);
+  textArea.value = text;
+  textArea.select();
+  document.execCommand('copy');
+  document.body.removeChild(textArea);
+};
